ta: add remote processor Trusted Application
Add a generic TA for a remote processor firmware management.
This TA should be defined as early TA to allow to kernel built-in
and remote processor firmware management in boot stages.

This change duplicates ELF header file for remoteproc TA to use
ELF support since ldelf library is not reachable from in-tree TAs.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I9c0f91d3b8d32d7a652837532e6406ce43b873ba