core: upgrade stm32_tamp driver
Adds full support for TAMP peripheral.

The tamp driver manages the anti-tamper detection circuit.
Sensitive data in the device are protected from external attack by the
anti-tamper detection circuit. There are internal tampers (linked to
internal states) and external tampers (linked to one or two external pins)
that may raise a tamper event if some event is detected.
External tamper can be configured for edge detection, or level detection
with or without filtering, or active tamper which increases the security
level by auto checking that the tamper pins are not externally opened or
shorted.

This driver provides also an API to configure backup registers permission
access and lock (and make usable) the SAES BHK key.

Signed-off-by: Nicolas Toromanoff <nicolas.toromanoff@foss.st.com>
Change-Id: I22cd52506081d9794efbf4306e26c419490ec7e4
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/232976
Tested-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>
Reviewed-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>