plat-stm32mp1: optimize xlat table allocation
Set MAX_XLAT_TABLES to the lowest possible value depending on
OP-TEE configuration when pager is enable.

Set 4kB translation to 4 tables is sufficient to map Core.
Note that this assumes client will not request mapping of
more than 2MByte at a time.

Decrease Core heap size from 64kByte to 48kByte in order to save
memory for the pager. This configuration is not mandated and
platform configuration can override this value.

Change-Id: I0b751576965d9fbb12a16a099bfec0c7188c0733
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>