drivers: stm32_risab: support RISAB driver
This driver implements the RISAB driver. Through RISAB registers, a
trusted compartment, or the compartiment to which the page
configuration has been delegated, assigns memory pages
to one or more security domains.

Each RISAB is dedicated to a memory and can cover 128KBytes of data,
separated in 32 pages of 4 KBytes, containing 8 blocks each.

It is possible to align a RISAB secure and privilege regions
allocations with an ARM Cortex M, which defines in its address space
configurable regions with a 256Bytes granularity. The configuration
would be 512Bytes block-based in order to align the two.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I04b2fb1666bfdd4a90ba84df534a7e76027231d1
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/257007
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>