pta: stm32mp: add new remoteproc pta
Add remoteproc pta skeleton inspired from the remoteproc_pta.c available
for the stm32mp1 platform.

As a first step this pta supports only the stm32mp2 relying on the
stm32_remoteproc driver.

In a second step it should also support the stm32mp1, and the legacy pta
of the stm32mp1 platform will be removed.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/264183
Change-Id: Ie77ad2c2dfed30f779f53df4bc0af2a4c8c32465