ta: remoteproc: refactor the signed format
Update to support the new signed format according to
the "script/sign_rproc_fw.py" signing tool.
Notice that the legacy format is no more supported.
This new format is based on a trailer chunk that contains
meta data in type-length-value format.

The platform TLV is not yet supported and will be added in
separate commits.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/265648
Change-Id: I360e50680d4093085805784029103022f4bd7272