drivers: stm32_remoteproc: rework boot addresses and trust zone management
Remove the management of the boot addresses and the trust zone
based on device tree.
Replace it by the stm32_rproc_set_boot_address API.
This API allow PTA to provide the Secure and non secure boot address
extracted from the platform TLV metadata.
If the secure boot address is set the Cortex-M trust-Zone is enabled.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/266708
Change-Id: Ic85de9d4981a6be984bead71baa857f75a80c33b