drivers: stm32_bsec: implement the get otp by phandle
Add a new interface stm32_bsec_find_otp_by_phandle() to retrieve an OTP
from a given phandle.

When the phandle is absent in the NVMEM node, layout_cell->phandle = 0 and
reference to this OTP with this API is not possible.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Change-Id: Iebec3bb7953c056ee72a00eca43b2ca14c796696
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/267530
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/281720
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/297445
Tested-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>