clk: stm32mp25: [TEMPORARY] put Cortex-M33 peripheral clock as critical
Force as critical the clocks used for the Cortex-M33 peripherals
locks to ensure that they are clocked for the Cortex-M33.
This is done to ensure that the flexgen channel is enabled.
The tim2 is not set a critical, indeed the source clock is not
the flexgen.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I53b601e2e088764e04290e9b9132ad645afc5004
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/295375
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Gatien CHEVALLIER <gatien.chevallier@st.com>