dt-bindings: describe st,protreg properties for reserved memories
The st,protreg describes the firewall hardware configuration
to apply for the region. This configuration will be applied by the
firewall controller. The macro used to describe the content of the
configuration varies. It depends on the type of the firewall controller

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I53e3990fc7f0382a00ac8f6195a68263fe08a82f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/313408
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>